Structure updated - Added Packages and Module close corporation
